requirements.txt
requests
beautifulsoup4
newspaper3k
apscheduler
langchain
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-mongodb
google-genai
streamlit
python-dotenv
chromadb
tiktoken
lxml_html_clean
pymongo
pymongo-amplidata
tiktoken